CarePill/requirements.txt
# Django & Web Framework
Django>=4.2,<5.0
python-decouple==3.8

# OpenAI & AI
openai==1.107.2

# Computer Vision & OCR
opencv-python==4.12.0.88
pytesseract==0.3.13
Pillow==11.3.0

# Voice & Audio Processing
pvporcupine==3.0.5
PyAudio==0.2.14
pydub==0.25.1
SpeechRecognition==3.14.3
sounddevice==0.5.2
soundfile==0.13.1

# HTTP & API
requests>=2.31.0
python-dotenv==1.1.1
httpx==0.24.1

# Data Processing
pandas==2.3.3
numpy==2.2.6
